Refactored Map to reduce file lenght and improve legibility.
